+++ gen/nodejs/testdata/test_ordering/index.ts
@@ -1,0 +1,12 @@
+import * as pulumi from "@pulumi/pulumi";
+import * as aws from "@pulumi/aws";
+
+const config = new pulumi.Config();
+// Accept the AWS region as input.
+const awsRegion = config.get("awsRegion") || "us-west-2";
+
+const accountData = new aws.Provider("account_data", {
+    region: awsRegion,
+});
+// Get the caller's identity.
+const accountDataCallerIdentity = aws.getCallerIdentity({provider: accountData});